remove header/footer from examples
rename
